Rename all usalt to microsalt
except the config keys
