+++ cg/meta/lims/__init__.py
@@ -1,1 +1,1 @@
-from .usalt import MicrosaltAPI
+from .microsalt import MicrosaltAPI
